Simplify the form handleChange
I applied the technique of using named properties for the key part of the object in square brackets, then can pull off the name and value from the e.target object and spread the prevNote. Important to note I used the function approach, to handle for the case note could be null and you cant spread a null object, if it is null, it creates and empty object, then title and description are spread into that object.